asterick.py
Use a for loop and range() function to display the following pattern in the terminal using a bunch of asterisks *:

*
* *
* * *
* * * *
* * * * *
# ... and so on

It should look like this but with 24 rows total. The first row should have one asterisk and the last row should have 24 asterisks.

Note: Make sure an empty space is between each asterisk.